preview: Back button on /profile too (shared component moved to components/profile)
